+++ src/com/milanesachan/DRPGTest1/bot/core/BattleCommandManager.java
@@ -1,6 +1,9 @@
 package com.milanesachan.DRPGTest1.bot.core;
 
-import com.milanesachan.DRPGTest1.bot.handlers.battle.SetBattleChannelHandler;
+import com.milanesachan.DRPGTest1.bot.entities.GuildParty;
+import com.milanesachan.DRPGTest1.bot.handlers.HandlerFilter;
+import com.milanesachan.DRPGTest1.bot.handlers.battle.*;
+import net.dv8tion.jda.api.entities.MessageChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
@@ -9,8 +12,7 @@ import java.util.ArrayList;
 
 public class BattleCommandManager extends ListenerAdapter {
     private static BattleCommandManager instance;
-    private String[] battleCommands = {"battle"};
-
+    private final ArrayList<GuildParty> parties;
 
     public static BattleCommandManager getInstance(){
         if(instance == null)
@@ -18,6 +20,9 @@ public class BattleCommandManager extends ListenerAdapter {
         return instance;
     }
 
+    private BattleCommandManager(){
+        parties = new ArrayList<>();
+    }
 
     @Override
     public void onMessageReceived(@Nonnull MessageReceivedEvent event) {
@@ -30,12 +35,81 @@ public class BattleCommandManager extends ListenerAdapter {
                     event.getGuild().getIdLong());
             h.handle();
         }
-        //First check if the received message is a battle command
-        //If it is, and it's not coming from a battle channel, print it
-        //If it is coming from a battle channel, continue
+        if(matchCommand(args[0], "battle")){
+            long guildID = event.getGuild().getIdLong();
+            long userID = event.getMember().getUser().getIdLong();
+            CreatePartyHandler h = new CreatePartyHandler(event.getChannel(), guildID, userID);
+
+            HandlerFilter filter = new HandlerFilter();
+            filter.setBattleChannelRequired(true);
+            filter.setCharacterRequired(true);
+            filter.setGuildMemberRequired(true);
+            filter.filterHandler(event, h, userID);
+        }else if(matchCommand(args[0], "invite")){
+            if(args.length < 2 ||
+            event.getMessage().getMentionedMembers().isEmpty()){
+                event.getChannel().sendMessage("Problem in command. Correct format is: '>invite <@User>'.").queue();
+            }else{
+                long invitedUserID = event.getMessage().getMentionedMembers().get(0).getIdLong();
+
+                PartyInviteHandler h = new PartyInviteHandler(event.getChannel(), event.getGuild().getIdLong(), invitedUserID);
+
+                HandlerFilter filter = new HandlerFilter();
+                filter.setBattleChannelRequired(true);
+                filter.setGuildMemberRequired(true);
+                filter.setCharacterRequired(true);
+                filter.filterHandler(event, h, invitedUserID);
+            }
+        }else if(matchCommand(args[0], "party")) {
+            long guildID = event.getGuild().getIdLong();
+
+            PartyHandler h = new PartyHandler(event.getChannel(), guildID);
+
+            HandlerFilter filter = new HandlerFilter();
+            filter.setBattleChannelRequired(true);
+            filter.filterHandler(event, h, 0);
+        }else if(matchCommand(args[0], "queue")){
+            long guildID = event.getGuild().getIdLong();
+
+            BattleQueueHandler h = new BattleQueueHandler(event.getChannel(), guildID);
+
+            HandlerFilter f = new HandlerFilter();
+            f.setBattleChannelRequired(true);
+            f.setGuildMemberRequired(true);
+            f.setPartyRequired(true);
+            f.setPartyMemberRequired(true);
+            f.filterHandler(event, h, event.getMember().getUser().getIdLong());
+        }
     }
 
     private boolean matchCommand(String comm, String test) {
         return comm.equalsIgnoreCase(DRPGBot.getInstance().getPrefix().concat(test));
+    }
+
+    public ArrayList<GuildParty> getParties(){
+        return parties;
+    }
+
+    public GuildParty getParty(long guildID){
+        for(GuildParty p : parties){
+            if(p.getGuildID() == guildID) return p;
+        }
+        return null;
+    }
+
+    public boolean removeParty(long guildID){
+        GuildParty toRemove = null;
+        for(GuildParty p : parties){
+            if(p.getGuildID() == guildID) toRemove = p;
+        }
+        if(toRemove != null) return parties.remove(toRemove);
+        else return false;
+    }
+
+    public boolean doesPartyExist(long guildID){
+        for(GuildParty p : parties){
+            if(p.getGuildID() == guildID) return true;
+        }
+        return false;
     }
 }