+++ src/com/milanesachan/DRPGTest1/commons/parameters/CharacterParameters.java
@@ -1,0 +1,5 @@
+package com.milanesachan.DRPGTest1.commons.parameters;
+
+public class CharacterParameters {
+    public static int initialMaxHealth = 500;
+}